requirements.txt
gunicorn==20.0.4
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.3
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
munch==2.5.0
requests==2.28.1
six==1.16.0
urllib3==1.26.10
Werkzeug==2.1.2
python-dotenv==0.20.0
